Added test for lexer. Rearranged files
